% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeeyedry.R
\name{write_cnt}
\alias{write_cnt}
\title{Write a region-boundary data frame to a standard EYEDRY .cnt file}
\usage{
write_cnt(df, filepath)
}
\arguments{
\item{df}{A data frame as returned by \code{\link{read_cnt}} or
\code{\link{extract_regions_from_script}}, containing
columns \code{item}, \code{cond}, \code{nreg},
\code{end_last}, and \code{start_R1 ... start_Rn}.}

\item{filepath}{File path where the .cnt file will be written.}
}
\description{
Write a region-boundary data frame to a standard EYEDRY .cnt file
}
